requirements.txt
tensorflow-cpu==2.6.0
streamlit==0.86.0
numpy==1.19.5